fix: color pallete design
