+++ src/common/_mixins.scss
@@ -1,21 +1,38 @@
 @mixin dropdown-arrow() {
-  content: "";
-  pointer-events: none;
-  display: block;
-  width: 0;
-  height: 0;
-  border-left: 3px solid transparent;
-  border-right: 3px solid transparent;
-  border-top: 5px solid;
-  margin-left: $grid-size-small;
+	content: "";
+	pointer-events: none;
+	display: block;
+	width: 0;
+	height: 0;
+	border-left: 3px solid transparent;
+	border-right: 3px solid transparent;
+	border-top: 5px solid;
+	margin-left: $grid-size-small;
 
-  // This gives the icon space on the right side consistent with the material
-  // icon standards.
-  margin-right: 2px;
+	// This gives the icon space on the right side consistent with the material
+	// icon standards.
+	margin-right: 2px;
 }
 
 // Formatting Buttons.
 @mixin formatting-button-style__hover {
-  color: $dark-gray-500;
-  box-shadow: inset 0 0 0 1px $dark-gray-500, inset 0 0 0 2px $white;
+	color: $dark-gray-500;
+	box-shadow: inset 0 0 0 1px $dark-gray-500, inset 0 0 0 2px $white;
+}
+
+@mixin reduce-motion($property: "") {
+	@if $property == "transition" {
+		@media (prefers-reduced-motion: reduce) {
+			transition-duration: 0s;
+		}
+	} @else if $property == "animation" {
+		@media (prefers-reduced-motion: reduce) {
+			animation-duration: 1ms;
+		}
+	} @else {
+		@media (prefers-reduced-motion: reduce) {
+			transition-duration: 0s;
+			animation-duration: 1ms;
+		}
+	}
 }
