+++ src/common/_variables.scss
@@ -12,3 +12,6 @@ $inserter-tabs-height: 36px;
 $block-toolbar-height: $block-controls-height + $border-width;
 $resize-handler-size: 15px;
 $resize-handler-container-size: $resize-handler-size + ($grid-size-small * 2); // Make the resize handle container larger so there's a larger grabbable area.
+
+// Buttons & UI Widgets
+$radius-round: 50%;
